--- a/.jenkins/jenkins_pr_gpu.sh
+++ b/.jenkins/jenkins_pr_gpu.sh
@@ -17,7 +17,7 @@ export LIBRARY_PATH=/usr/local/cuda/lib64:$LIBRARY_PATH
 
 echo "===== Testing gpuarray backend"
 
-GPUARRAY_CONFIG="Release"
+GPUARRAY_CONFIG="Debug"
 DEVICE=cuda
 LIBDIR=${WORKSPACE}/local
 
@@ -62,6 +62,6 @@ python -c "import theano.gpuarray; theano.gpuarray.use('${DEVICE}')" || { echo '
 THEANO_GPUARRAY_TESTS="theano/gpuarray/tests \
                        theano/scan_module/tests/test_scan.py:T_Scan_Gpuarray \
                        theano/scan_module/tests/test_scan_checkpoints.py:TestScanCheckpoint.test_memory"
-FLAGS="init_gpu_device=$DEVICE,gpuarray.preallocate=1000,mode=DEBUG_MODE,on_opt_error=raise,on_shape_error=raise,cmodule.age_thresh_use=604800,base_compiledir=$BASECOMPILEDIR"
+FLAGS="init_gpu_device=$DEVICE,gpuarray.preallocate=1000,mode=FAST_RUN,on_opt_error=raise,on_shape_error=raise,cmodule.age_thresh_use=604800,base_compiledir=$BASECOMPILEDIR"
 FLAGS=${FLAGS},magma.enabled=true # Enable magma GPU library
 THEANO_FLAGS=${FLAGS} time nosetests --with-xunit --xunit-file=theanogpuarray_tests.xml ${THEANO_GPUARRAY_TESTS}