--- a/theano/gpuarray/dnn.py
+++ b/theano/gpuarray/dnn.py
@@ -3765,6 +3765,9 @@ def local_dnn_reduction(node):
                 scal = 'norm1'
             else:
                 return
+        elif (isinstance(node.op.scalar_op, theano.scalar.basic.Maximum) and
+                isinstance(node.op.pre_scalar_op, theano.scalar.basic.Abs)):
+            scal = 'absmax'
         else:
             return