… theano/gpuarray/tests/test_dnn.py