@abergeron 2017-10-19
I'm wondering why we don't just disable the elemwise fusion.