@nouiz 2017-10-19
It will raise the peak memory usage and will be much slower then with this PR.

old code:
- V100: 0.136s
- TITAN Xp: 0.127s

new code:
v100: 0.000235s
Titan Xp: 0.00035s

speed up:
- V100 578x
- Titan Xp 361x