@notoraptor 2017-10-20
One Jenkins failure: http://earlgrey.iro.umontreal.ca:8080/job/Theano_PR/961/testReport/junit/theano.gpuarray.tests/test_opt/test_local_gpu_elemwise_careduce/

Fixed: 4f2862a