@notoraptor 2017-10-20
@nouiz Updated, I added optimization for `MAX(ABS(X))`.

I cannot currently add optimization for average (`MEAN(X)`) because theano `mean()` function produces a graph with a sum divided iteratively by axes (something like `x.sum() / shape[0] / shape[1] / ...`). So I don't know if it is a good idea to parse such case.

Mean function is here: https://github.com/Theano/Theano/blob/master/theano/tensor/basic.py#L3272