@notoraptor 2017-10-23
I am not able to reproduce the problem (tested on: cuda 8 + cudnn 7002/7003, then cuda 9 + cudnn 7003).