@nouiz 2017-10-23
jenkins failure: http://earlgrey.iro.umontreal.ca:8080/job/Theano_PR/968/

jenkins test this

Maybe it was a random error, so I start the test again, just in case.