@slefrancois 2017-10-23
Jenkins tests were stuck so I killed and restarted manually.