@notoraptor 2017-10-23
In fact, error can be reproduced with `floatX=float16`: http://earlgrey.iro.umontreal.ca:8080/job/Theano_PR/972/console
```
[Theano_GPU]         // rest of function is handled by one warp
[Theano_GPU]         if (threadNum < warpSize) {
[Theano_GPU]             //round up all the partial sums into the first `warpSize` elements
[Theano_GPU]             for (int i = threadNum + warpSize; i < threadCount; i += warpSize)
[Theano_GPU]             {
[Theano_GPU]                 myresult = ((buf[i])>(myresult)? (buf[i]): ((myresult)>=(buf[i])? (myresult): nan("")));
[Theano_GPU]             }
[Theano_GPU]             buf[threadNum] = myresult;
[Theano_GPU]         }
[Theano_GPU]         __syncthreads();
[Theano_GPU]         for (unsigned int _n = warpSize / 2; _n > 0; _n /= 2) {
[Theano_GPU]             if (threadNum < _n && threadNum + _n < threadCount)
[Theano_GPU]             buf[threadNum] = ((buf[threadNum+_n])>(buf[threadNum])? (buf[threadNum+_n]): ((buf[threadNum])>=(buf[threadNum+_n])? (buf[threadNum]): nan("")));
[Theano_GPU]             __syncthreads();
[Theano_GPU]         }
[Theano_GPU]         if (threadNum == 0) {
[Theano_GPU]           Z[0] = (buf[0]);
[Theano_GPU]         }
[Theano_GPU]         
[Theano_GPU]             }
[Theano_GPU]             
[Theano_GPU] NVRTC compile log::
[Theano_GPU] default_program(11): error: more than one instance of overloaded function "fabs" matches the argument list:
[Theano_GPU]             function "fabs(double)"
[Theano_GPU]             function "fabs(float)"
[Theano_GPU]             argument types are: (const short)
```

It seems a call to `fabs()` receives a `short` but can not then choose between `fabs(double)` and `fabs(float)`. It seems related to how the kernel code is generated for pre scalar op. My fix in last commit seems to resolve the problem: https://github.com/Theano/Theano/pull/6491/commits/5e80e5fa46a0e47159c59052ae1e84579b9e155e

@nouiz @abergeron @lamblin