@nouiz 2017-10-24
travis fail with wrong result errors.