@notoraptor 2017-10-25
Still jenkins errors: http://earlgrey.iro.umontreal.ca:8080/job/Theano_PR/977/

I retry with libgpuarray compiled in `Release` mode.