@notoraptor 2017-10-25
Oh, that's strange: no more error when libgpuarray is compiled in Release mode: http://earlgrey.iro.umontreal.ca:8080/job/Theano_PR/980/

Is it something expectable ? @abergeron @nouiz @lamblin 

I guess no.