@nouiz 2017-10-25
I do not expect changing from Debug compilation to Release compilation of libgpuarray to change the results. I'll review in detail this PR. Can you make a new PR that set again the Debug compilation of libgpuarray?