@lamblin 2017-10-25
When libgpuarray is compiled in Debug, it also adds an offset to each memory allocation.
If incorrect results get produced, it could mean that the offset is not correctly passed. Since the offset is zero for most arrays, compiling in Release could work most of the time, but still be incorrect in some cases.