@nouiz 2017-10-25
@notoraptor can you try to reproduce those error locally by building libgpuarray in Debug?

if you can reproduce them, try with Theano mode=DebugMode in addtion to the above. This could help identify which op have the problem.