@notoraptor 2017-10-25
@nouiz Yes I am able to reproduce it on Ubuntu (by compiling libgpuarray in Debug mode). I will try theano debug mode tomorrow.

Note that DNN errors are fixed by @abergeron 's commit : http://earlgrey.iro.umontreal.ca:8080/job/Theano_PR/977/