@notoraptor 2017-10-26
(bad jenkins link above, this is the last one in the other PR): http://earlgrey.iro.umontreal.ca:8080/job/Theano_PR/984/