@notoraptor 2017-10-26
This PR can be closed, as it continues with #6496 .